+++ nextgisweb/vector_layer/model.py
@@ -121,6 +121,11 @@ class VectorLayer(Base, Resource, SpatialLayerMixin, LayerFieldsMixin):
     def _tablename(self):
         return 'layer_%s' % self.tbl_uuid
 
+    def _drop_table(self, connection):
+        tableinfo = TableInfo.from_layer(self)
+        tableinfo.setup_metadata(self._tablename)
+        tableinfo.metadata.drop_all(bind=connection)
+
     def from_ogr(self, filename, *, layername=None):
         ds = read_dataset(filename)
         layer = ds.GetLayerByName(layername) if layername is not None else ds.GetLayer(0)
@@ -349,9 +354,7 @@ event.listen(
 # Drop data table on vector layer deletion
 @event.listens_for(VectorLayer, 'before_delete')
 def drop_verctor_layer_table(mapper, connection, target):
-    tableinfo = TableInfo.from_layer(target)
-    tableinfo.setup_metadata(target._tablename)
-    tableinfo.metadata.drop_all(bind=connection)
+    target._drop_table(connection)
 
 
 class _source_attr(SP):
@@ -413,7 +416,8 @@ class _source_attr(SP):
 
     def setter(self, srlzr, value):
         if srlzr.obj.id is not None:
-            raise VE("Source parameter does not apply to update vector layer.")
+            srlzr.obj._drop_table(DBSession.connection())
+            srlzr.obj.tbl_uuid = uuid.uuid4().hex
 
         datafile, metafile = env.file_upload.get_filename(value['id'])
 
@@ -473,13 +477,20 @@ class _geometry_type_attr(SP):
 
     def setter(self, srlzr, value):
         if value not in GEOM_TYPE.enum:
-            raise VE(_("Unsupported geometry type."))
+            raise VE(message=_("Unsupported geometry type."))
 
         if srlzr.obj.id is None:
             srlzr.obj.geometry_type = value
 
         elif srlzr.obj.geometry_type != value:
-            raise VE(_("Geometry type for existing resource can't be changed."))
+            raise VE(message=_("Geometry type for existing resource can't be changed."))
+
+
+class _delete_all_features_attr(SP):
+
+    def setter(self, srlzr, value):
+        if value:
+            srlzr.obj.feature_delete_all()
 
 
 P_DSS_READ = DataStructureScope.read
@@ -494,7 +505,9 @@ class VectorLayerSerializer(Serializer):
 
     srs = SR(read=P_DSS_READ, write=P_DSS_WRITE)
 
-    source = _source_attr(read=None, write=P_DS_WRITE)
+    source = _source_attr(write=P_DS_WRITE)
 
     geometry_type = _geometry_type_attr(read=P_DSS_READ, write=P_DSS_WRITE)
-    fields = _fields_attr(read=None, write=P_DS_WRITE)
+    fields = _fields_attr(write=P_DS_WRITE)
+
+    delete_all_features = _delete_all_features_attr(write=P_DS_WRITE)